+++ lib/asset/datamanager_validation_test.go
@@ -31,6 +31,7 @@ func TestDataManagerValidate(t *testing.T) {
 			Description:    validAddressable,
 			Opener:         validAddressable,
 			Type:           "test",
+			LogsPermission: validPermissions,
 		}, false},
 		"valid": {&NewDataManager{
 			Key:            "4c67ad88-309a-48b4-8bc4-c2e2c1a87a83",
@@ -39,6 +40,7 @@ func TestDataManagerValidate(t *testing.T) {
 			Description:    validAddressable,
 			Opener:         validAddressable,
 			Type:           "test",
+			LogsPermission: validPermissions,
 		}, true},
 	}