+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- add a `logs_permission` field to the Dataset asset (#459)
+
 ## [0.4.0] - 2022-01-05
 
 ### Added